moved-moody-camel-to-external
